Make new file for stats endpoint
