+++ listenbrainz/webserver/__init__.py
@@ -240,6 +240,7 @@ def _register_blueprints(app):
     from listenbrainz.webserver.views.profile import profile_bp
     from listenbrainz.webserver.views.follow import follow_bp
     from listenbrainz.webserver.views.follow_api import follow_api_bp
+    from listenbrainz.webserver.views.stats_api import stats_api_bp
     from listenbrainz.webserver.views.status_api import status_api_bp
     app.register_blueprint(index_bp)
     app.register_blueprint(login_bp, url_prefix='/login')
@@ -248,5 +249,6 @@ def _register_blueprints(app):
     app.register_blueprint(follow_bp, url_prefix='/follow')
     app.register_blueprint(api_bp, url_prefix=API_PREFIX)
     app.register_blueprint(follow_api_bp, url_prefix=API_PREFIX+'/follow')
+    app.register_blueprint(stats_api_bp, url_prefix=API_PREFIX+'/stats')
     app.register_blueprint(status_api_bp, url_prefix=API_PREFIX+'/status')
     app.register_blueprint(api_bp_compat)
